[Refactor] ディレクトリ構造を再設計
