+++ dispander/__init__.py
@@ -1,1 +1,5 @@
-from dispander.module import *
+from dispander.cogs.expand import ExpandDiscordMessageFromUrlCog
+
+
+async def setup(bot):
+    await bot.add_cog(ExpandDiscordMessageFromUrlCog(bot))
